Implement function to load home page on button click
We also implemented a function that removes all the children of co-
ntent. Now, this way, we have essentially reloaded that specific
aspect of the page. I will say however that I don't know whether
reducing text content to an empty string to remove the children
elements is the most efficient approach, but I hope it is.